+++ ray-marching-lib/sphere/sphere.h
@@ -6,37 +6,37 @@
 typedef struct Sphere
 {
     Vec3 position;
-    double radius;
+    float radius;
 } Sphere;
 
-static inline Sphere sphere_create(double radius, double x, double y, double z)
+static inline Sphere sphere_create(float radius, float x, float y, float z)
 {
     Vec3 vec = {x, y, z};
     Sphere new_sphere = {vec, radius};
     return new_sphere;
 }
 
-static inline double distance_to_sphere(const Vec3 *origin, const Sphere *sphere)
+static inline float distance_to_sphere(const Vec3 *origin, const Sphere *sphere)
 {
-    double dx = origin->x - sphere->position.x;
-    double dy = origin->y - sphere->position.y;
-    double dz = origin->z - sphere->position.z;
+    float dx = origin->x - sphere->position.x;
+    float dy = origin->y - sphere->position.y;
+    float dz = origin->z - sphere->position.z;
 
     return sqrt(dx * dx + dy * dy + dz * dz) - sphere->radius;
 }
 
-static inline double distance_to_closest_sphere(const Vec3 *origin, const Sphere spheres[], int numSpheres)
+static inline float distance_to_closest_sphere(const Vec3 *origin, const Sphere spheres[], int numSpheres)
 {
-    double minDist = INFINITY;
+    float minDist = INFINITY;
 
     for (int i = 0; i < numSpheres; i++)
     {
-        minDist = fmin(distance_to_sphere(origin, &spheres[i]), minDist);
+        minDist = fminf(distance_to_sphere(origin, &spheres[i]), minDist);
     }
 
     return minDist;
 
-    // double dist1 = distance_to_sphere(origin, &spheres[0]);
-    // double dist2 = distance_to_sphere(origin, &spheres[1]);
+    // float dist1 = distance_to_sphere(origin, &spheres[0]);
+    // float dist2 = distance_to_sphere(origin, &spheres[1]);
     // return smoothMin(dist1, dist2, 1.5);
 }
