+++ ray-marching-lib/utils/utils.h
@@ -1,6 +1,6 @@
 #pragma once
 
-static inline Vec3 spherical_coordinates_to_vector(double theta, double phi, double radius)
+static inline Vec3 spherical_coordinates_to_vector(float theta, float phi, float radius)
 {
     Vec3 vector;
 
@@ -11,8 +11,8 @@ static inline Vec3 spherical_coordinates_to_vector(double theta, double phi, dou
     return vector;
 }
 
-static inline double smoothMin(double d1, double d2, double k)
+static inline float smoothMin(float d1, float d2, float k)
 {
-    double h = fmax(k - fabs(d1 - d2), 0.0) / k;
-    return fmin(d1, d2) - h * h * h * k * (1.0 / 6.0);
+    float h = fmax(k - fabsf(d1 - d2), 0.0) / k;
+    return fminf(d1, d2) - h * h * h * k * (1.0 / 6.0);
 }
